requirements.txt
streamlit
playwright
google-genai
fpdf2